+++ CHANGELOG.md
@@ -20,11 +20,16 @@ Note that since we don't clearly distinguish between a public and private interf
 - Fix aromatic ring detection not accounting for hybridization
 - Add axis param to camera spin/rock animation
 - Fix SSAO half/quarter resolution textures for multi-scale
-- Camera improvements 
+- Camera improvements
   - Add the option to approximate "least obstructed direction" when focusing camera, accessibe via `PluginContext.managers.camera.focusLoci` with `optimizeDirection` option
   - Add `CameraFocusOptions.zoomOut` option that zooms out to to make the entire scene visible before focusing on the target
   - Add easing support in camera transtion
 - Non-covalent interactions: water bridge support
+- Add OBJ format support
+    - Positions, normals, faces
+    - Groups from usemtl directive
+    - Vertex color extension
+    - Sideload MTL files (diffuse color only)
 - Download Structure From AlphaFoldDB allows IDs with version suffix (version is ignored)
 - Add `loadUrl` method and GET params to Viewer app